Move most of the admin bar component to a new AdminBarApp.
