+++ assets/js/googlesitekit-adminbar.js
@@ -19,94 +19,24 @@
 /**
  * External dependencies
  */
-import './modules';
 
 /**
  * WordPress dependencies
  */
-import { render, Fragment, useCallback } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
+import { render } from '@wordpress/element';
 
 /**
  * Internal dependencies.
  */
-import Data from 'googlesitekit-data';
 import {
-	decodeHTMLEntity,
 	loadTranslations,
-	trackEvent,
 } from './util';
-import Link from './components/Link';
-import AdminbarModules from './components/adminbar/adminbar-modules';
 import Root from './components/root';
-import { STORE_NAME as CORE_SITE } from './googlesitekit/datastore/site/constants';
-const { useSelect } = Data;
+import AdminBarApp from './components/adminbar/AdminBarApp';
+import './modules';
 
 export const GoogleSitekitAdminbar = () => {
-	const currentEntityURL = useSelect( ( select ) => select( CORE_SITE ).getCurrentEntityURL() );
-	const currentEntityTitle = useSelect( ( select ) => select( CORE_SITE ).getCurrentEntityTitle() );
-	const detailsURL = useSelect( ( select ) => select( CORE_SITE ).getAdminURL( 'googlesitekit-dashboard', { permaLink: currentEntityURL } ) );
-
-	const onMoreDetailsClick = useCallback( async () => {
-		await trackEvent( 'admin_bar', 'post_details_click' );
-		document.location.assign( detailsURL );
-	}, [ detailsURL ] );
-
-	if ( ! detailsURL || ! currentEntityURL ) {
-		return null;
-	}
-
-	return (
-		<Fragment>
-			<div className="mdc-layout-grid">
-				<div className="mdc-layout-grid__inner">
-					<div className="
-						mdc-layout-grid__cell
-						mdc-layout-grid__cell--span-3
-						mdc-layout-grid__cell--align-middle
-					">
-						<div className="googlesitekit-adminbar__subtitle">{ __( 'Stats for', 'google-site-kit' ) }</div>
-						<div className="googlesitekit-adminbar__title">
-							{ currentEntityTitle
-								? decodeHTMLEntity( currentEntityTitle )
-								: currentEntityURL
-							}
-						</div>
-					</div>
-					<div className="
-						mdc-layout-grid__cell
-						mdc-layout-grid__cell--span-8-tablet
-						mdc-layout-grid__cell--span-7-desktop
-						mdc-layout-grid__cell--align-middle
-					">
-						<div className="mdc-layout-grid__inner">
-							<AdminbarModules />
-						</div>
-					</div>
-					<div className="
-						mdc-layout-grid__cell
-						mdc-layout-grid__cell--span-2
-						mdc-layout-grid__cell--align-middle
-					">
-						<Link
-							className="googlesitekit-adminbar__link"
-							href="#"
-							onClick={ onMoreDetailsClick }
-						>
-							{ __( 'More details', 'google-site-kit' ) }
-						</Link>
-					</div>
-				</div>
-			</div>
-			<Link
-				className="googlesitekit-adminbar__link googlesitekit-adminbar__link--mobile"
-				href="#"
-				onClick={ onMoreDetailsClick }
-			>
-				{ __( 'More details', 'google-site-kit' ) }
-			</Link>
-		</Fragment>
-	);
+	return <AdminBarApp />;
 };
 
 // Initialize the whole adminbar app.
